Move fusion fixup rewrites to their own pass
These patterns shouldn't be in conv-to-gemm, both because they'll be
applicable to a GEMM by itself and because they're only used when
fusion is enabled.
